Warn project for outdated dependencies
Projects using oparent as parent should have information about outdated
dependencies. This is done at validation phase.

A (template of a) list of rules is provided to ignore certain versions,
a more importantly to ignore all versions of an artifact but one.
This might be use to warn projects not using the needed dependency
of an artifact when preparing for an specific release of ONAP.

Current rule file is commented as it breaks it's not available on the
git repository.

Change-Id: I30bbbe5d14d8f297f046700963cc5a238e41901f
Issue-id:INT-121
Signed-off-by: Anaël Closson <ac2550@intl.att.com>