Revert display of maven artifacts to be updated
The operation was added to the default lifecycle and consumes a lot of
time on each build on the LF infrastructure, more or less depending on
the amount of dependencies to be checked.

A project with ~60 dependencies can see an increase of more than an hour
on its verify job with this plugin activated.

This patch completely reverts the versions-maven-plugin while we're
looking either for a way to speed up this plugin or we find an alternative.

Change-Id: I53451d4f3e27e5c8fd07df46b05c1e313af48633
Issue-Id: INT-121
Signed-off-by: Anaël Closson <ac2550@intl.att.com>