Optional execution off versions-maven-plugin
The maven plugin used for displaying the list of artifacts that can be
updated consumes a lot of time on each job execution on the LF infrastructure.

The execution should be optional and can be configured by assigning it
on a specific maven phase when needed using the following variable:
 versions.plugin.phase

By default, it's assigned to a non existing phase to disable it.

Change-Id: I30d38a8458e6e855ff27cf3c31678cbc373e59c9
Issue-Id: INT-121
Signed-off-by: Anaël Closson <ac2550@intl.att.com>