Update patch set 1
Patch Set 1:

As it stands, the release-version job will fail when it sets the release version on the oparent root pom since the oparent-odlparent artifact doesn't inherit its version from the root pom.  We may need to set up a separate set of Jenkins jobs and release the oparent-odlparent artifact separately from the rest of oparent.  Any other suggestions?

If we go with the separate Jenkins jobs approach, then we should remove the inclusion of oparent-odlparent as a module in the root pom.

Patch-set: 1