Update patch set 1
Patch Set 1:

I'm not sure why the release-version job would fail?  That job uses the version plugin (not release plugin) to reset the version number and executes "update-child-modules" ... so the <version> tag in child modules will also be reset.

If there's an issue I'm missing that we can work around using a separate Jenkins job, that's also fine.  I'll go with whatever approach you'd prefer. 

FYI - I actually tried just copying oparent to ccsdk/parent, but when I did that, that didn't work (it couldn't find the stylecheck pom.xml for one thing).  I was able to a local build of this change, then a local build of ccsdk/parent using this one as its parent, and then sli/core using my updated parent, and that seemed to work, but of course that's a local build.  Mileage often varies on Jenkins

Patch-set: 1
Label: Code-Review=0