OParent children do not find checkstyle
The checkstyle plugin cannot be found by projects inheriting by OParent.
When maven executes the check-license execution, he finds the
org.onap.oparent:checkstyle:jar:1.0.0 as dependency and tries to retreive it.
The repository section doesn't apply to plugin dependencies and so
the org.onap.oparent:checkstyle:jar cannot be found.

By adding it as dependency of the org.onap.oparent:oparent project,
the artifact is downloaded using the project repositories and is already
available when the plugin dependencies are checked.

This might be due to a maven bug.

Change-Id: I9eabea660b5fe8ceacc7299914e64ccccd7c2850
Issue-Id: INT-128
Signed-off-by: Anaël Closson <ac2550@intl.att.com>