Fix checkstyle build and xml
The XML for onap-java-code style has a bug identified by Eclipse. The
maxLineLength field is not allowed.

In addition, not sure how any jobs work already, the sourceDirectory
appends and extra "src/main/java" to the path so some builds don't
always find the correct source files to generate warnings. The plugin
is a bit flaky and sometimes seems to find the java files, while other
times it does not. So some checkstyle warnings are not generated.

Change-Id: Ia07065a0d0df1429643680cb5bd70b5ec13bebde
Issue-ID: TSC-71
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>