Update patch set 2
Patch Set 2: Code-Review+1

At this point, this is the only dependency upgrade needed for the Policy Framework. If other PTL's find that something in this pom.xml needs to be upgraded, they should open a ticket and get it in ASAP so we can release this artifact. Or someone could look at the CLM report for this report and do appropriate upgrades.

Patch-set: 2
Reviewer: Gerrit User 57 <57@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1