update version of checkstyle for latest google style and jdk 8
Upgrade to use the latest google style (plus onap style).
This latest style adds rules for jdk 8 source level.

Upgrade the version of checkstyle to 8.24.  The latest google
style rules are compatible with 8.24 and later.

Upgrade checkstyle maven plugin to 3.1.0 (from 2.17).

This jdk8 version of checkstyle is active only when the build-time
java runtime is jdk8 or later, otherwise, the original checkstyle
is active.

Issue-ID: INT-1271
Change-Id: Ib190b39efc5a6028ab4874301e6a813f0a36b138
Signed-off-by: John J. Franey <john.franey@att.com>