Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I think Pam mentioned that master has some issues.. 
> Although, I just remembered that Sonar jobs track only master :(
> Unless we are planning for tech teams repos master branch to point to this future release from elalto oparent

Yeah we only run sonar on master branch - because if you run for multiple branches they overwrite the results. Last run wins - so probably don't need this for elalto branch.

No issues with master since it is a SNAPSHOT that very few projects are pointing to. Its being upgraded to support JDK 11, so there *may* be issues with some plugins needing to be upgraded. But I don't think so for sonar plugin. I think you are safe.

Patch-set: 2