Update patch set 2
Patch Set 2: Code-Review-1

That's a lot of new committers, was there a vote initiated? And results presented to the TSC? In other words, there would be a line added to the bottom of the file recording that. Or did the by-laws change for committers?

Patch-set: 2
Reviewer: Gerrit User 57 <57@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1