Remove self references in oparent
This commit removes self references in oparent. THese self references
mean that the version of an import in oparent is set to the valie of
project.version, which is the value of project.version in the project
importing oparent, not the actual version of oparent. This causes all
sorts of side effects.

The down side is that users of oparent will now have to explicitly
import the oparent dependencies, doing somethling like:

  <dependencyManagement>
    <dependencies>
      <dependency>
        <groupId>org.onap.oparent</groupId>
        <artifactId>dependencies</artifactId>
        <version><Your oparent version></version>
        <type>pom</type>
        <scope>import</scope>
      </dependency>
    </dependencies>
  </dependencyManagement>

Issue-ID: POLICY-4482
Change-Id: I573bffb0656209413fafc3800c3bcbdb616beecd
Signed-off-by: liamfallon <liam.fallon@est.tech>